test: move jar parser tests into parse_test.go
